+++ .gitignore
@@ -1,3 +1,6 @@
+# Agent generated by frida-compile
+/_agent.js
+
 # Logs
 logs
 *.log
