+++ agent/hook/cccrypt.js
