+++ agent/hook/index.js
