+++ agent/jailbreak.js
