+++ agent/lib/libc.js
