+++ agent/lib/nsdict.js
