--- agent/passionfruit.plist
@@ -1,1 +1,0 @@
-{ Filter = { Bundles = ( "com.apple.Security" ); }; }
