+++ lib/rpc.js
@@ -1,7 +1,9 @@
 const { Transform } = require('stream')
 
 const socketStream = require('socket.io-stream')
-const fridaLoad = require('frida-load')
+const { promisify } = require('util')
+const fs = require('fs')
+const readFile = promisify(fs.readFile)
 
 const Cache = require('./cache')
 const { retry } = require('./utils')
@@ -26,7 +28,7 @@ module.exports = class RpcHandler {
 
   async load() {
     const { socket } = this
-    const source = await fridaLoad(require.resolve('../frida'))
+    const source = await readFile(require.resolve('../_agent'), 'utf8')
     const script = await this.session.createScript(source)
 
     await script.load()
@@ -86,7 +88,6 @@ module.exports = class RpcHandler {
 
       'info',
       'userDefaults',
-      'lsof',
       'imports',
 
       'ls',
@@ -171,10 +172,6 @@ module.exports = class RpcHandler {
 
   async userDefaults() {
     return this.agent.userDefaults()
-  }
-
-  async lsof() {
-    return this.agent.lsof()
   }
 
   async imports(data) {
