+++ package.json
@@ -3,10 +3,31 @@
   "version": "0.0.2",
   "description": "Inject to iOS App",
   "main": "app.js",
+  "browserify": {
+    "transform": [
+      [
+        "babelify",
+        {
+          "presets": [
+            [
+              "es2015",
+              {
+                "loose": true
+              }
+            ]
+          ],
+          "plugins": [
+            "transform-runtime"
+          ]
+        }
+      ]
+    ]
+  },
   "scripts": {
     "dev": "cross-env NODE_ENV=development nodemon .",
     "start": "cross-env NODE_ENV=production node .",
-    "build": "pushd gui && npm run build && popd && pushd agent && make package install && popd",
+    "build": "pushd gui && npm run build && popd && frida-compile agent -o _agent.js",
+    "watch": "frida-compile agent -o _agent.js -w",
     "postinstall": "cd gui && (yarn || npm install)"
   },
   "repository": {
@@ -27,8 +48,6 @@
     "base64-js": "^1.2.1",
     "eslint": "^4.8.0",
     "frida": "^10.6.3",
-    "frida-fs": "^1.1.1",
-    "frida-load": "^1.0.0",
     "koa": "^2.3.0",
     "koa-bodyparser": "^4.2.0",
     "koa-compress": "^2.0.0",
@@ -36,9 +55,7 @@
     "koa-logger": "^3.0.1",
     "koa-router": "^7.2.1",
     "koa-send": "^4.1.1",
-    "mz": "^2.6.0",
     "nodemon": "^1.12.1",
-    "objection": "^0.8.7",
     "plist": "^2.1.0",
     "promise.prototype.finally": "^3.0.0",
     "socket.io": "^2.0.3",
@@ -46,8 +63,14 @@
     "sqlite3": "^3.1.9"
   },
   "devDependencies": {
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-runtime": "^6.26.0",
+    "babelify": "^7.3.0",
     "cross-env": "^5.0.5",
     "eslint-config-airbnb-base": "^12.0.2",
-    "eslint-plugin-import": "^2.7.0"
+    "eslint-plugin-import": "^2.7.0",
+    "frida-compile": "^5.1.0",
+    "macho": "^1.2.1"
   }
 }
